unexport funcs and move to better place
